% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/nullcat.R
\name{nullcat_methods}
\alias{nullcat_methods}
\title{Supported nullcat methods}
\usage{
nullcat_methods()

nullcat_methods()
}
\value{
A character vector of method names.

Character vector of available method names.
}
\description{
Return the character vector of supported categorical randomization methods.
}
\examples{
nullcat_methods()

}
