% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{trace_nullcat}
\alias{trace_nullcat}
\title{Trace diagnostics for nullcat mixing}
\usage{
trace_nullcat(
  x,
  method = "curvecat",
  max_iter = 1000L,
  n_runs = 3,
  stat = "kappa",
  plot = FALSE
)
}
\arguments{
\item{x}{A matrix containing integer or factor values.}

\item{method}{Character specifying algorithm to use. See \code{nullcat()} for details.}

\item{max_iter}{Number of successive matrix updates to apply. Default is 1000.}

\item{n_runs}{Number of independent runs to perform. Default is 3.}

\item{stat}{Output trace statistic to compute for each iteration. Either:
\itemize{
   \item \code{"kappa"} (the default) Cohen's kappa between the original `x` and
   the randomized matrix `x_rand`
   \item \code{"prop_equal"} proportion of cells in which `x == x_rand`
   \item any function taking (`x`, `x_rand`) and returning a numeric scalar
}}

\item{plot}{Logical; if TRUE, plot the trace. Default is FALSE.}
}
\value{
A matrix of trace stat values, with a row for each iteration
   and a column for each run.
}
\description{
Applies \code{nullcat()} to a categorical matrix, recording a summary statistic
at each iteration to help assess mixing on a given dataset.
}
