% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvecat.R
\name{curvecat}
\alias{curvecat}
\title{Categorical curveball randomization (curvecat)}
\usage{
curvecat(x, n_iter = 1000L, output = c("category", "index"), seed = NULL)
}
\arguments{
\item{x}{A matrix of categorical data, encoded as integers.
Values should represent category or stratum membership for each cell.}

\item{n_iter}{Number of iterations. Default is 1000. Larger values yield
more thorough mixing. Ignored for non-sequential methods. Minimum
burn-in times can be estimated with \link{suggest_n_iter}.}

\item{output}{Character indicating type of result to return:
\itemize{
  \item \code{"category"} (default) returns randomized matrix
  \item \code{"index"} returns an index matrix describing where original
       entries moved.
}}

\item{seed}{Integer used to seed random number generator, for reproducibility.}
}
\value{
A matrix of the same dimensions as \code{x}, either randomized
  categorical values (when \code{output = "category"}) or an integer index
  matrix describing the permutation of entries (when \code{output = "index"}).
}
\description{
Categorical generalization of the binary curveball algorithm
(Strona et al.) to matrices of categorical data. This function
is a convenience wrapper around [nullcat()] with
`method = "curvecat"`.
}
\details{
The curvecat algorithm randomizes a categorical matrix while keeping
the category multisets of each row and column fixed. In other words,
the permuted matrix has the same set of integer values in every row
and every column as the original matrix, but they are permuted. It
operates on pairs of rows at a time, grouping differing entries by
unordered category pairs and redistributing the orientation of those
pairs while preservingn the multiset of categories within each row.
When there are only two categories, `curvecat()` reduces to the
behavior of the original binary curveball algorithm applied to a 0/1
matrix.
}
\references{
Strona, G., Nappo, D., Boccacci, F., Fattorini, S., & San-Miguel-Ayanz, J.
(2014). A fast and unbiased procedure to randomize ecological binary
matrices with fixed row and column totals. \emph{Nature Communications}, 5,
4114.
}
\seealso{
[nullcat()], [nullcat_methods()]
}
