% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nullcat_methods}
\alias{nullcat_methods}
\title{Supported nullcat methods}
\usage{
nullcat_methods()
}
\value{
A character vector of method names.
}
\description{
Return the character vector of supported categorical randomization methods.
}
